+++ tests/Integration/Framework/Container/Tagged/ValidatorInterface.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GacelaTest\Integration\Framework\Container\Tagged;
+
+interface ValidatorInterface
+{
+    public function name(): string;
+}